+++ src/extension.ts
@@ -1,29 +1,65 @@
-// The module 'vscode' contains the VS Code extensibility API
-// Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode'
 
-// this method is called when your extension is activated
-// your extension is activated the very first time the command is executed
+interface Config {
+  include: string
+  exclude: string
+}
+
+const intersection = <T>(a: Set<T>, b: Set<T>): Set<T> => {
+  const results = new Set<T>()
+  for (const value of a) {
+    if (b.has(value)) {
+      results.add(value)
+    }
+  }
+  return results
+}
+
+const countFiles = async (
+  include: string,
+  exclude: string
+): Promise<number> => {
+  let first = true
+  let results = new Set<string>()
+  const includePatterns = include.split(',')
+  const excludePatterns = exclude.split(',')
+  for (const include of includePatterns) {
+    for (const exclude of excludePatterns) {
+      const res = await (
+        await vscode.workspace.findFiles(include, exclude)
+      ).map((uri) => uri.fsPath)
+
+      if (first) {
+        results = new Set(res)
+        first = false
+        continue
+      }
+
+      results = intersection(results, new Set(res))
+    }
+  }
+  return results.size
+}
+
+const getConfig = (): Config => {
+  const config = vscode.workspace.getConfiguration('file-count')
+  return {
+    include: config.get('include') || '**',
+    exclude: config.get('exclude') || 'node_modules/',
+  }
+}
+
 export function activate(context: vscode.ExtensionContext): void {
-  // Use the console to output diagnostic information (console.log) and errors (console.error)
-  // This line of code will only be executed once when your extension is activated
-  console.log('Congratulations, your extension "file-count" is now active!')
-
-  // The command has been defined in the package.json file
-  // Now provide the implementation of the command with registerCommand
-  // The commandId parameter must match the command field in package.json
   const disposable = vscode.commands.registerCommand(
-    'file-count.helloWorld',
-    () => {
-      // The code you place here will be executed every time your command is executed
-
-      // Display a message box to the user
-      vscode.window.showInformationMessage('Hello World from File Count!')
+    'file-count.showNumOfFiles',
+    async () => {
+      const config = getConfig()
+      const count = await countFiles(config.include, config.exclude)
+      vscode.window.showInformationMessage(count.toString())
     }
   )
 
   context.subscriptions.push(disposable)
 }
 
-// this method is called when your extension is deactivated
 export function deactivate(): void {}