+++ CHANGELOG.md
@@ -6,4 +6,6 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+### Added
+- "Show # of Files" command
+- Include/Exclude configuration options